Adds rollup
- Concatenates all defintions into one file for "typings"
- Uses a two pass, maybe we'll use rollup permanently
- Migrates cds-hooks types directly into the module
