--- cds-hooks.d.ts
@@ -1,2 +1,0 @@
-// https://github.com/Automate-Medical/sero/issues/50
-export * from './dist/cds-hooks';
