+++ rollup.config.js
@@ -1,0 +1,11 @@
+import dts from "rollup-plugin-dts";
+
+const config = [
+  {
+    input: "./dist/types/index.d.ts",
+    output: [{ file: "dist/index.d.ts", format: "es" }],
+    plugins: [dts()],
+  },
+];
+
+export default config;
