+++ src/index.ts
@@ -2,7 +2,7 @@
  * @module sero/sero
  */
 
-export { default as Rest } from "./rest";
-export { default as CDSHooks } from "./cds-hooks";
+export * as Rest from "./rest";
+export * as CDSHooks from "./cds-hooks";
 export { default as Config } from "./config";
-export { default as Http } from "./http";
+export * as Http from "./http";
